extracts buildFormState logic from endpoint for reuse
