+++ packages/next/src/exports/utilities.ts
@@ -1,6 +1,5 @@
 export { addDataAndFileToRequest } from '../utilities/addDataAndFileToRequest.js'
 export { addLocalesToRequestFromData, sanitizeLocales } from '../utilities/addLocalesToRequest.js'
-export { traverseFields } from '../utilities/buildFieldSchemaMap/traverseFields.js'
 export { createPayloadRequest } from '../utilities/createPayloadRequest.js'
 export { getNextRequestI18n } from '../utilities/getNextRequestI18n.js'
 export { getPayloadHMR, reload } from '../utilities/getPayloadHMR.js'
