+++ packages/ui/src/forms/Form/index.tsx
@@ -69,6 +69,8 @@ export const Form: React.FC<FormProps> = (props) => {
     waitForAutocomplete,
   } = props
 
+  console.log('Form', disabledFromProps)
+
   const method = 'method' in props ? props?.method : undefined
 
   const router = useRouter()
