+++ packages/ui/src/forms/buildStateFromSchema/calculateDefaultValues/iterateFields.ts
@@ -1,4 +1,5 @@
-import type { Data, Field, PayloadRequestWithData, TabAsField } from 'payload/types'
+import type { User } from 'payload/auth'
+import type { Data, Field, TabAsField } from 'payload/types'
 
 import { defaultValuePromise } from './promise.js'
 
@@ -6,16 +7,18 @@ type Args<T> = {
   data: T
   fields: (Field | TabAsField)[]
   id?: number | string
-  req: PayloadRequestWithData
+  locale: string | undefined
   siblingData: Data
+  user: User
 }
 
 export const iterateFields = async <T>({
   id,
   data,
   fields,
-  req,
+  locale,
   siblingData,
+  user,
 }: Args<T>): Promise<void> => {
   const promises = []
   fields.forEach((field) => {
@@ -24,8 +27,9 @@ export const iterateFields = async <T>({
         id,
         data,
         field,
-        req,
+        locale,
         siblingData,
+        user,
       }),
     )
   })
