+++ packages/ui/src/forms/buildStateFromSchema/calculateDefaultValues/promise.ts
@@ -1,12 +1,7 @@
+import type { User } from 'payload/auth'
 import type { Data } from 'payload/types'
 
-import {
-  type Field,
-  type PayloadRequestWithData,
-  type TabAsField,
-  fieldAffectsData,
-  tabHasName,
-} from 'payload/types'
+import { type Field, type TabAsField, fieldAffectsData, tabHasName } from 'payload/types'
 import { getDefaultValue } from 'payload/utilities'
 
 import { iterateFields } from './iterateFields.js'
@@ -15,8 +10,9 @@ type Args<T> = {
   data: T
   field: Field | TabAsField
   id?: number | string
-  req: PayloadRequestWithData
+  locale: string | undefined
   siblingData: Data
+  user: User
 }
 
 // TODO: Make this works for rich text subfields
@@ -24,8 +20,9 @@ export const defaultValuePromise = async <T>({
   id,
   data,
   field,
-  req,
+  locale,
   siblingData,
+  user,
 }: Args<T>): Promise<void> => {
   if (fieldAffectsData(field)) {
     if (
@@ -34,8 +31,8 @@ export const defaultValuePromise = async <T>({
     ) {
       siblingData[field.name] = await getDefaultValue({
         defaultValue: field.defaultValue,
-        locale: req.locale,
-        user: req.user,
+        locale,
+        user,
         value: siblingData[field.name],
       })
     }
@@ -52,8 +49,9 @@ export const defaultValuePromise = async <T>({
         id,
         data,
         fields: field.fields,
-        req,
+        locale,
         siblingData: groupData,
+        user,
       })
 
       break
@@ -70,8 +68,9 @@ export const defaultValuePromise = async <T>({
               id,
               data,
               fields: field.fields,
-              req,
+              locale,
               siblingData: row,
+              user,
             }),
           )
         })
@@ -97,8 +96,9 @@ export const defaultValuePromise = async <T>({
                 id,
                 data,
                 fields: block.fields,
-                req,
+                locale,
                 siblingData: row,
+                user,
               }),
             )
           }
@@ -115,8 +115,9 @@ export const defaultValuePromise = async <T>({
         id,
         data,
         fields: field.fields,
-        req,
+        locale,
         siblingData,
+        user,
       })
 
       break
@@ -136,8 +137,9 @@ export const defaultValuePromise = async <T>({
         id,
         data,
         fields: field.fields,
-        req,
+        locale,
         siblingData: tabSiblingData,
+        user,
       })
 
       break
@@ -148,8 +150,9 @@ export const defaultValuePromise = async <T>({
         id,
         data,
         fields: field.tabs.map((tab) => ({ ...tab, type: 'tab' })),
-        req,
+        locale,
         siblingData,
+        user,
       })
 
       break
